@treeowl 2023-01-30
Could you please rebase this and update the before/after benchmarks (since `transposeG` has changed)?