@meooow25 2023-01-30
Yes it's pretty much the same.