@meooow25 2023-01-30
Sure, happy to implement any suggested improvements to make the code cleaner.